Adding full pre-release version comparison support.
I was toying with this library and I had the need to do a little more accurate
pre-release comparisons. I saw the TODO and decided to bite if off. I started by
adding the values at the bottom of secion 11 here: http://semver.org to the
tests and change
https://github.com/amerine/go-semver/blob/29b30448e895b3a00419003f027196bf17ca2a58/semver/semver.go#L99
to actually check versionB.

This resulted in some failures that I could work against. The solution here
needs to check for ints and strings based on the precedence rules that semver
prefers. I was hoping for a cleaner way of doing that, but nothing I threw at
the problem was quicker than what I'm using here (manual type conversions and
leaning on the success/failure of that to decide where to proceed.)