Semver: handle PreRelease disparity according to (11)
With the help of an additional test statement in this patch, fixed
a bug related to the end of (11) on semver.org: if two pre-releases
are otherwise equal, the one with more elements is larger.

Before the fix, the additional test condition showed:
    --- FAIL: TestCompare (0.00s)
    semver_test.go:89: 1.2.3-a should be less than 1.2.3-a.b
    semver_test.go:89: 1.0.0-beta should be less than 1.0.0-beta.2
    semver_test.go:89: 1.0.0-alpha should be less than 1.0.0-alpha.1

The last 2 are from (11). After the fix, the tests now pass.